requirements.txt
click>=8.1.0,<9.0.0
feedparser>=6.0.10,<7.0.0
requests>=2.31.0,<3.0.0
pydub>=0.25.1,<1.0.0
openai-whisper>=20231117
openai>=1.14.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
yt-dlp>=2024.1.0